+++ packages/admin-ui/src/experimental.ts
@@ -8,6 +8,15 @@ import {
   useDatePickerState as experimental_useDatePickerState,
 } from './date-picker'
 import {
+  useFilterMultipleState as experimental_useFilterMultipleState,
+  FilterMultiple as experimental_FilterMultiple,
+  useFilterState as experimental_useFilterState,
+  Filter as experimental_Filter,
+  FilterGroup as experimental_FilterGroup,
+  useFilterGroupState as experimental_useFilterGroupState,
+} from './filters'
+
+import {
   I18nProvider as experimental_I18nProvider,
   useDateFormatter as experimental_useDateFormatter,
 } from './i18n'
@@ -16,6 +25,14 @@ import {
   ComboboxPopover as experimental_ComboboxPopover,
   useComboboxState as experimental_useComboboxState,
 } from './combobox'
+import {
+  Inline as experimental_Inline,
+  InlineProps as experimental_InlineProps,
+} from './inline'
+import {
+  Bleed as experimental_Bleed,
+  BleedProps as experimental_BleedProps,
+} from './bleed'
 
 export {
   experimental_DatePickerField,
@@ -30,4 +47,14 @@ export {
   experimental_ComboboxField,
   experimental_ComboboxPopover,
   experimental_useComboboxState,
+  experimental_FilterGroup,
+  experimental_FilterMultiple,
+  experimental_Filter,
+  experimental_useFilterGroupState,
+  experimental_useFilterMultipleState,
+  experimental_useFilterState,
+  experimental_Inline,
+  experimental_InlineProps,
+  experimental_Bleed,
+  experimental_BleedProps,
 }