+++ packages/admin-ui/src/filters/filter-group.state.tsx
@@ -1,0 +1,15 @@
+import type { GenericFilterStateReturn } from './base-filter.state'
+
+export function useFilterGroupState(props: UseFilterGroupStateReturn) {
+  const { filterStates } = props
+
+  const onClear = () => {
+    filterStates.forEach((state) => state.onClear())
+  }
+
+  return { onClear }
+}
+
+interface UseFilterGroupStateReturn {
+  filterStates: GenericFilterStateReturn[]
+}