+++ packages/admin-ui-hooks/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/vtex/admin-ui/compare/@vtex/admin-ui-hooks@0.2.1...@vtex/admin-ui-hooks@0.2.2) (2022-03-18)
+
+**Note:** Version bump only for package @vtex/admin-ui-hooks
+
 ## [0.2.1](https://github.com/vtex/admin-ui/compare/@vtex/admin-ui-hooks@0.2.0...@vtex/admin-ui-hooks@0.2.1) (2022-03-16)
 
 **Note:** Version bump only for package @vtex/admin-ui-hooks