+++ packages/formik/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.17](https://github.com/vtex/admin-ui/compare/@vtex/admin-formik@0.4.16...@vtex/admin-formik@0.4.17) (2022-03-18)
+
+**Note:** Version bump only for package @vtex/admin-formik
+
+## [0.4.16](https://github.com/vtex/admin-ui/compare/@vtex/admin-formik@0.4.15...@vtex/admin-formik@0.4.16) (2022-03-18)
+
+**Note:** Version bump only for package @vtex/admin-formik
+
+## [0.4.15](https://github.com/vtex/admin-ui/compare/@vtex/admin-formik@0.4.14...@vtex/admin-formik@0.4.15) (2022-03-17)
+
+**Note:** Version bump only for package @vtex/admin-formik
+
+## [0.4.14](https://github.com/vtex/admin-ui/compare/@vtex/admin-formik@0.4.13...@vtex/admin-formik@0.4.14) (2022-03-17)
+
+**Note:** Version bump only for package @vtex/admin-formik
+
 ## [0.4.13](https://github.com/vtex/admin-ui/compare/@vtex/admin-formik@0.4.12...@vtex/admin-formik@0.4.13) (2022-03-16)
 
 **Note:** Version bump only for package @vtex/admin-formik